requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
httpx==0.25.0
pyjwt[crypto]==2.10.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
minio==7.2.0
redis==5.0.7
requests==2.32.3
prometheus-client==0.17.0